Extracts common functionality to its own class
Fixes: SIRI-243
